Implement Flow Matching with CPU-based EMA and optimizations
Major updates:
- Fix tensor stacking issues in FMDataCollator with proper tokenizer padding
- Implement CPU-based EMA model to save GPU memory
- Add conditional EMA updates based on configurable intervals
- Fix dtype compatibility issues in flow matching scheduler
- Update model configuration: 30 layers, vocab_size 30, max_length 1022
- Add comprehensive dataset/cache exclusions to .gitignore
- Include EMA testing utilities

+++ model/backbone/flow_matching_scheduler.py
@@ -167,7 +167,7 @@ class DiscreteAbsorbingFlowMatchingScheduler:
         seq_indices = torch.arange(clean_tokens.size(1)).unsqueeze(0)
         
         # Set target probabilities for clean tokens at corrupted positions
-        target_distribution[batch_indices, seq_indices, clean_tokens] = corruption_mask.float()
+        target_distribution[batch_indices, seq_indices, clean_tokens] = corruption_mask.to(target_distribution.dtype)
         
         # Convert logits to log probabilities
         log_probs = F.log_softmax(model_logits, dim=-1)
